+++ studio/src/app/components/landing/app-landing-content/app-landing-content.scss
@@ -1,0 +1,85 @@
+app-landing-content {
+  div.introducing, div.remote {
+    background: white;
+  }
+
+  div.introducing, div.audience, div.remote {
+    section {
+      display: grid;
+      grid-template-columns: 1fr 1fr;
+      grid-template-rows: auto;
+      grid-gap: 64px;
+      align-items: center;
+
+      padding-top: 64px;
+
+      iframe {
+        box-shadow: 8px 8px 16px rgba(0, 0, 0, .12);
+        border-radius: 8px;
+      }
+
+      @media (max-width: 768px) {
+        display: flex;
+      }
+    }
+  }
+
+  div.introducing, div.remote {
+    section {
+      @media (max-width: 768px) {
+        flex-direction: column-reverse;
+      }
+    }
+  }
+
+  div.audience {
+    section {
+      @media (max-width: 768px) {
+        flex-direction: column;
+      }
+    }
+  }
+
+  @media (max-width: 768px) {
+    .wider-devices {
+      display: none;
+    }
+  }
+
+  div.audience {
+
+    div.deck {
+      position: relative;
+      display: block;
+
+      width: 100%;
+      height: 320px;
+
+      --slide-width: 100%;
+      --slide-height: 240px;
+
+      box-shadow: 8px 8px 16px rgba(0, 0, 0, .12);
+      border-radius: 8px;
+
+      background: white;
+
+      deckgo-slide-poll {
+        --slide-padding-top: 8px;
+        --slide-padding-start: 16px;
+        --slide-padding-end: 16px;
+        --slide-padding-bottom: 8px;
+
+        --slide-poll-grid-column-gap: 0;
+
+        --slide-poll-justify-content: flex-start;
+
+        p[slot="awaiting-votes"] {
+          text-align: center;
+          width: 75%;
+          border: 0;
+        }
+      }
+    }
+
+  }
+}
