+++ studio/src/app/components/landing/app-landing/app-landing.scss
@@ -30,176 +30,6 @@ app-landing {
       height: calc(100vh - 56px);
 
       background: var(--ion-color-primary);
-
-      deckgo-deck {
-        --deckgo-highlight-code-font-size: 125%;
-
-        h1 {
-          font-weight: 600;
-          margin-bottom: 0;
-        }
-
-        h2, h3 {
-          font-weight: 500;
-        }
-
-        h1, h2, h3 {
-          margin-top: 4px;
-        }
-
-        *[slot]:not([slot="background"]) {
-          z-index: 2;
-        }
-
-        div[slot="background"] {
-          width: calc(100% + 2px);
-          height: 100%;
-          z-index: 1;
-
-          > img.wave {
-            position: absolute;
-            bottom: 0;
-            left: -1px;
-          }
-
-          deckgo-lazy-img {
-            --deckgo-lazy-img-height: 100%;
-            --deckgo-lazy-img-min-width: 100%;
-            --deckgo-lazy-img-object-fit: cover;
-          }
-
-          button.action {
-            border: none;
-            outline: none;
-            font: inherit;
-            color: inherit;
-            background: none;
-            padding: 0;
-
-            font-size: 3.8rem;
-
-            display: flex;
-            justify-content: center;
-            align-items: center;
-
-            cursor: pointer;
-            pointer-events: all;
-
-            transition: font-size 300ms;
-
-            &:hover, &:active {
-              font-size: 4.6rem;
-            }
-
-            &.next, &.prev {
-              position: absolute;
-              top: 50%;
-            }
-
-            &.next {
-              right: 0;
-              transform: translate(0, -50%);
-            }
-
-            &.prev {
-              left: 0;
-              transform: translate(0, -50%);
-            }
-
-            @media (max-width: 768px) {
-              &.next, &.prev {
-                top: unset;
-                bottom: 4rem;
-                transform: translate(0, 0);
-              }
-            }
-          }
-        }
-      }
     }
-  }
-
-  div.introducing, div.remote {
-    background: white;
-  }
-
-  div.introducing, div.audience, div.remote {
-    section {
-      display: grid;
-      grid-template-columns: 1fr 1fr;
-      grid-template-rows: auto;
-      grid-gap: 64px;
-      align-items: center;
-
-      padding-top: 64px;
-
-      iframe {
-        box-shadow: 8px 8px 16px rgba(0,0,0,.12);
-        border-radius: 8px;
-      }
-
-      @media (max-width: 768px) {
-        display: flex;
-      }
-    }
-  }
-
-  div.introducing, div.remote {
-    section {
-      @media (max-width: 768px) {
-        flex-direction: column-reverse;
-      }
-    }
-  }
-
-  div.audience {
-    section {
-      @media (max-width: 768px) {
-        flex-direction: column;
-      }
-    }
-  }
-
-  @media (max-width: 768px) {
-    .wider-devices {
-      display: none;
-    }
-  }
-
-  div.audience {
-
-    div.deck {
-      position: relative;
-      display: block;
-
-      width: 100%;
-      height: 320px;
-
-      --slide-width: 100%;
-      --slide-height: 240px;
-
-      box-shadow: 8px 8px 16px rgba(0,0,0,.12);
-      border-radius: 8px;
-
-      background: white;
-
-      deckgo-slide-poll {
-        --slide-padding-top: 8px;
-        --slide-padding-start: 16px;
-        --slide-padding-end: 16px;
-        --slide-padding-bottom: 8px;
-
-        --slide-poll-grid-column-gap: 0;
-
-        --slide-poll-justify-content: flex-start;
-
-      p[slot="awaiting-votes"] {
-          text-align: center;
-          width: 75%;
-          border: 0;
-        }
-      }
-    }
-
   }
 }
